# openapi.yaml
...
openapi: 3.1.0
info:
  title: За столом — API
  version: 1.1.0
  description: 'Контракт server/index.mjs и server/store.mjs. JSON без envelope: списки
    — массивы; ошибки — {error:string}. Все деньги JSON API в целых копейках. JSON-тела
    <=65536 байт. POST /api/auth/*: максимум 180 запросов за 10 минут на IP; TRUST_PROXY=true
    доверяет X-Real-IP только за настроенным reverse proxy. Неизвестные JSON-поля
    не назначают роли, цену или scope. Отсутствуют отмена, reopen, оплата
    и интеграции POS. Публичный server URL — назначение развёртывания, не заявление
    о его готовности.'
servers:
- url: https://mail.lonelycraft.ru/banquet
  description: Назначение production через nginx, который удаляет /banquet
- url: http://localhost:3100
  description: Локальный Docker Compose
- url: http://localhost:3000
  description: Прямой Node.js backend
security:
- SessionBearer: []
paths:
...
  "/api/config":
    get:
      operationId: config
      summary: Публичная конфигурация интерфейса
      description: ''
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Config"
        '500':
          "$ref": "#/components/responses/500"
      security: []
  "/api/auth/max":
    post:
      operationId: loginMax
      summary: Обмен подписанных данных MAX на сессию
      description: HMAC-SHA256 по алгоритму MAX; auth_date не старше часа и не более
        30 секунд в будущем. MAX user ID из RESTAURANT_ADMIN_IDS получает restaurant,
        остальные organizer. Live-сессия 7 суток; не более 10 активных сессий на пользователя.
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Session"
        '400':
          "$ref": "#/components/responses/400"
        '401':
          "$ref": "#/components/responses/401"
        '413':
          "$ref": "#/components/responses/413"
        '415':
          "$ref": "#/components/responses/415"
        '429':
          "$ref": "#/components/responses/429"
        '503':
          "$ref": "#/components/responses/503"
        '500':
          "$ref": "#/components/responses/500"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/MaxLogin"
      security: []
  "/api/auth/demo":
    post:
      operationId: loginDemo
      summary: Создать демо, сменить роль или войти по демо-приглашению
      description: Без sandbox создаётся отдельный scope с синтетическими данными.
        С sandbox выбирается роль в своём демо. inviteCode без sandbox создаёт только
        guest в scope приглашения, игнорируя запрошенные привилегии; live-приглашение
        даёт 403. Демо-сессия 24 часа, scope хранится 7 суток. Не более MAX_DEMO_SPACES
        (по умолчанию 1000) demo scope. DEMO_ENABLED=false также отзывает существующие
        демо-сессии и скрывает публичные демо-приглашения.
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DemoSession"
        '400':
          "$ref": "#/components/responses/400"
        '401':
          "$ref": "#/components/responses/401"
        '403':
          "$ref": "#/components/responses/403"
        '404':
          "$ref": "#/components/responses/404"
        '409':
          "$ref": "#/components/responses/409"
        '413':
          "$ref": "#/components/responses/413"
        '415':
          "$ref": "#/components/responses/415"
        '429':
          "$ref": "#/components/responses/429"
        '500':
          "$ref": "#/components/responses/500"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/DemoLogin"
      security: []
  "/api/max/webhook":
    post:
      operationId: maxWebhook
      summary: Принять событие MAX и ответить ссылкой Mini App
      description: Проверка секрета до чтения тела. Persistent claim перед ответом
        обеспечивает не более одной попытки welcome на событие. Тайм-аут MAX API 8
        секунд; ошибка доставки после claim подтверждается 200 без автоматической
        повторной отправки.
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/WebhookAck"
        '400':
          "$ref": "#/components/responses/400"
        '401':
          "$ref": "#/components/responses/401"
        '413':
          "$ref": "#/components/responses/413"
        '415':
          "$ref": "#/components/responses/415"
        '503':
          "$ref": "#/components/responses/503"
        '500':
          "$ref": "#/components/responses/500"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/MaxUpdate"
      security:
      - MaxWebhookSecret: []
...
  "/api/me":
    get:
      operationId: me
      summary: Текущий пользователь
      description: ''
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/User"
        '401':
          "$ref": "#/components/responses/401"
        '500':
          "$ref": "#/components/responses/500"
  "/api/me/phone":
    put:
      operationId: verifyPhone
      summary: Привязать подтверждённый номер телефона MAX к текущей сессии
      description: Тело — результат window.WebApp.requestContact(). Сервер проверяет HMAC-SHA256 токеном бота, свежесть authDate и userId из подписанной MAX-сессии; присланный клиентом userId не принимается. Номер нормализуется в международный формат без плюса. Демо-сессии запрещены.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/MaxContact"
      responses:
        '200':
          description: Обновлённый профиль
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/User"
        '400':
          "$ref": "#/components/responses/400"
        '401':
          "$ref": "#/components/responses/401"
        '403':
          "$ref": "#/components/responses/403"
        '409':
          "$ref": "#/components/responses/409"
  "/api/me/notifications":
    put:
      operationId: setNotifications
      summary: Включить или выключить уведомления в MAX
      description: Доставка возможна только после начала личного диалога с ботом. По умолчанию выключено; bot_stopped отключает фактическую отправку.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
      responses:
        '200':
          description: Обновлённый профиль
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/User"
        '400':
          "$ref": "#/components/responses/400"
        '401':
          "$ref": "#/components/responses/401"
        '403':
          "$ref": "#/components/responses/403"
  "/api/media":
    post:
      operationId: uploadPhoto
      summary: Загрузить фото для блюда или напитка
      description: Организатор или ресторан. Бинарное тело до 3 МиБ, проверяется сигнатура формата. Возвращённый photoUrl передаётся при создании или правке позиции меню.
      requestBody:
        required: true
        content:
          image/jpeg:
            schema: { type: string, format: binary }
          image/png:
            schema: { type: string, format: binary }
          image/webp:
            schema: { type: string, format: binary }
      responses:
        '201':
          description: Путь к загруженному фото
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PhotoUpload"
        '401':
          "$ref": "#/components/responses/401"
        '403':
          "$ref": "#/components/responses/403"
        '413':
          "$ref": "#/components/responses/413"
        '415':
          "$ref": "#/components/responses/415"
  "/api/media/{fileName}":
    get:
      operationId: getPhoto
      summary: Получить публичное фото блюда
      security: []
      parameters:
      - "$ref": "#/components/parameters/fileName"
      responses:
        '200':
          description: JPEG, PNG или WebP
        '404':
          "$ref": "#/components/responses/404"
  "/api/restaurants":
    get:
      operationId: restaurants
      summary: Рестораны и текущий каталог своего scope
      description: Все роли видят каталог только своего scope. sampleMenu обозначает
        пример каталога и сбрасывается при первой его правке.
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Restaurant"
        '401':
          "$ref": "#/components/responses/401"
        '500':
          "$ref": "#/components/responses/500"
  "/api/events":
    get:
      operationId: events
      summary: Доступные текущему пользователю банкеты
      description: Владелец, участник или ресторанная роль в том же scope. Сортировка
        created_at DESC; без пагинации.
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Event"
        '401':
          "$ref": "#/components/responses/401"
        '500':
          "$ref": "#/components/responses/500"
    post:
      operationId: createEvent
      summary: Создать банкет и снимок доступного меню
      description: Organizer или restaurant. Создание события и event_menu в транзакции. Не
        более 30 событий на демо-организатора или 500 на live-организатора.
      responses:
        '201':
          description: Успешно
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Event"
        '400':
          "$ref": "#/components/responses/400"
        '401':
          "$ref": "#/components/responses/401"
        '403':
          "$ref": "#/components/responses/403"
        '409':
          "$ref": "#/components/responses/409"
        '413':
          "$ref": "#/components/responses/413"
        '415':
          "$ref": "#/components/responses/415"
        '429':
          "$ref": "#/components/responses/429"
        '500':
          "$ref": "#/components/responses/500"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/EventCreate"
  "/api/events/{eventId}":
    get:
      operationId: event
      summary: Банкет, персональный выбор и разрешённая сводка
      description: ''
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/EventDetail"
        '401':
          "$ref": "#/components/responses/401"
        '404':
          "$ref": "#/components/responses/404"
        '500':
          "$ref": "#/components/responses/500"
      parameters:
      - "$ref": "#/components/parameters/eventId"
    patch:
      operationId: editEvent
      summary: Изменить дату, дедлайн, название, число гостей или бюджет
      description: Владелец или ресторан, пока заказ collecting. expectedRevision обязателен; при конкурирующем изменении возвращается 409. Число гостей не может быть меньше длины списка приглашённых.
      parameters:
      - "$ref": "#/components/parameters/eventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/EventPatch"
      responses:
        '200':
          description: Обновлённый банкет
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Event"
        '400':
          "$ref": "#/components/responses/400"
        '401':
          "$ref": "#/components/responses/401"
        '403':
          "$ref": "#/components/responses/403"
        '409':
          "$ref": "#/components/responses/409"
...
  "/api/events/{eventId}/selection":
    put:
      operationId: saveSelection
      summary: Полностью заменить собственный выбор
      description: До дедлайна, только collecting. В live нужен свой подтверждённый
        номер из списка приглашённых; в демо владелец присоединяется автоматически.
        Обновление транзакционное, revision увеличивается,
        цена берётся из event_menu.
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/SelectionSaved"
        '400':
          "$ref": "#/components/responses/400"
        '401':
          "$ref": "#/components/responses/401"
        '403':
          "$ref": "#/components/responses/403"
        '404':
          "$ref": "#/components/responses/404"
        '409':
          "$ref": "#/components/responses/409"
        '413':
          "$ref": "#/components/responses/413"
        '415':
          "$ref": "#/components/responses/415"
        '500':
          "$ref": "#/components/responses/500"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SelectionPut"
      parameters:
      - "$ref": "#/components/parameters/eventId"
...
  "/api/events/{eventId}/export.csv":
    get:
      operationId: exportCsv
      summary: Скачать утверждённый заказ для кухни
      description: 'Только owner или restaurant того же scope, статус approved. UTF-8
        BOM, разделитель ;, CRLF. Секции: банкет/дата, суммарные количества/суммы,
        позиции по гостям с пожеланиями. CSV суммы в рублях с 2 знаками; формулы в
        текстовых полях нейтрализуются.'
      responses:
        '200':
          description: CSV-файл
          headers:
            Content-Disposition:
              schema:
                type: string
              example: attachment; filename="banquet-kitchen.csv"
          content:
            text/csv:
              schema:
                type: string
        '401':
          "$ref": "#/components/responses/401"
        '403':
          "$ref": "#/components/responses/403"
        '404':
          "$ref": "#/components/responses/404"
        '409':
          "$ref": "#/components/responses/409"
        '500':
          "$ref": "#/components/responses/500"
      parameters:
      - "$ref": "#/components/parameters/eventId"
  "/api/events/{eventId}/export-link":
    post:
      operationId: createExportLink
      summary: Создать одноразовую ссылку CSV для MAX Bridge
      description: Bearer owner/restaurant, только approved. Тело не требуется. TTL60
        секунд, одна действующая ссылка на user/event; новая заменяет старую. Передать
        url в WebApp.downloadFile(url, filename) по пользовательскому клику.
      parameters:
      - "$ref": "#/components/parameters/eventId"
      responses:
        '200':
          description: Одноразовая ссылка
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExportLink"
        '401':
          "$ref": "#/components/responses/401"
        '403':
          "$ref": "#/components/responses/403"
        '404':
          "$ref": "#/components/responses/404"
        '409':
          "$ref": "#/components/responses/409"
        '500':
          "$ref": "#/components/responses/500"
  "/api/downloads/{downloadToken}":
    get:
      operationId: consumeExportLink
      summary: Однократно скачать CSV по короткоживущей ссылке
      description: 'Без Bearer: token в пути является capability. Проверяются срок,
        действующая роль владельца, scope, approved и revision. Ссылка расходуется
        атомарно; неизвестная, использованная, истёкшая или отозванная возвращает404.
        Cache-Control private,no-store; Referrer-Policy no-referrer.'
      security: []
      parameters:
      - "$ref": "#/components/parameters/downloadToken"
      responses:
        '200':
          description: CSV-файл
          headers:
            Content-Disposition:
              schema:
                type: string
              example: attachment; filename="banquet-kitchen.csv"
          content:
            text/csv:
              schema:
                type: string
        '404':
          "$ref": "#/components/responses/404"
        '500':
          "$ref": "#/components/responses/500"
components:
  securitySchemes:
    SessionBearer:
      type: http
      scheme: bearer
      description: Токен приложения из /api/auth/max или /api/auth/demo, не MAX_BOT_TOKEN.
    MaxWebhookSecret:
      type: apiKey
      in: header
      name: X-Max-Bot-Api-Secret
  parameters:
...
    fileName:
      name: fileName
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-f0-9-]{36}\\.(jpg|png|webp)$"
  schemas:
...
    Config:
      type: object
      properties:
        demoEnabled:
          type: boolean
        maxConfigured:
          type: boolean
          description: Наличие MAX_BOT_TOKEN; не доказывает доступность webhook, модерацию
            или правильность токена.
        botUsername:
          type: string
      required:
      - demoEnabled
      - maxConfigured
      - botUsername
...
    Session:
      type: object
      properties:
        token:
          type: string
          description: Bearer-секрет приложения. Хранится в БД только в SHA-256; не
            токен MAX.
        user:
          "$ref": "#/components/schemas/User"
      required:
      - token
      - user
...
    MaxLogin:
      type: object
      properties:
        initData:
          type: string
          minLength: 1
          maxLength: 16384
          description: Исходная строка window.WebApp.initData без повторной JSON-сериализации
            вложенного user.
      required:
      - initData
...
  responses:
    '400':
      description: Некорректные данные, JSON или отсутствие expectedRevision.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    '401':
      description: Отсутствует/истекла сессия, неверная MAX-подпись, sandbox либо
        webhook secret.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
...
    '413':
      description: JSON-тело превышает 65536 байт.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    '415':
      description: 'Ожидается Content-Type: application/json.'
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    '429':
      description: Лимит входов, демо-пространств, банкетов либо блюд превышен.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    '500':
      description: Внутренняя ошибка; технические детали не возвращаются клиенту.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    '503':
      description: MAX не настроен либо хранилище дедупликации временно недоступно.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"